taher_last_version

+++ lib/profile.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:confirm_dialog/confirm_dialog.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -9,6 +10,7 @@ import 'package:gojo/auth.dart';
 import 'package:gojo/generated/l10n.dart';
 import 'package:gojo/loginPage.dart';
 import 'package:gojo/my_profile.dart';
+import 'package:gojo/term_of_use.dart';
 
 class profile extends StatefulWidget {
   profile({super.key, required this.change});
@@ -47,7 +49,7 @@ class _profileState extends State<profile> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        backgroundColor: Color.fromARGB(255, 232, 228, 214),
+        backgroundColor: Color.fromARGB(255, 255, 138, 138),
         body: ListView(
           shrinkWrap: true,
           children: [
@@ -59,7 +61,7 @@ class _profileState extends State<profile> {
                     "${S.of(context).welcome} $_name ",
                     style: TextStyle(
                         fontSize: 50,
-                        color: const Color.fromARGB(255, 222, 95, 92)),
+                        color: Color.fromARGB(255, 248, 158, 158)),
                   );
                 } else {
                   return CircularProgressIndicator();
@@ -69,14 +71,17 @@ class _profileState extends State<profile> {
             Text(
               S.of(context).editProfile,
             ),
-            GestureDetector(
-              onTap: () {
-                print(Data);
-                Navigator.of(context).push(MaterialPageRoute(
-                    builder: (context) =>
-                        edit_profile(Data: Data!, email: _email!)));
-              },
+            Container(
+              margin: EdgeInsets.only(left: 15, right: 15, top: 10),
               child: ListTile(
+                  onTap: () {
+                    print(Data);
+                    Navigator.of(context).push(MaterialPageRoute(
+                        builder: (context) =>
+                            edit_profile(Data: Data!, email: _email!)));
+                  },
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(100)),
                   leading: Icon(
                     Icons.account_box_rounded,
                   ),
@@ -85,18 +90,23 @@ class _profileState extends State<profile> {
                   )),
             ),
             Text(S.of(context).accountSettings),
-            GestureDetector(
-              onTap: () {
-                FirebaseAuth.instance.currentUser!.delete();
-                Navigator.of(context).pushAndRemoveUntil(
-                    MaterialPageRoute(
-                      builder: (context) => SignInPage(
-                        change: widget.change,
-                      ),
-                    ),
-                    (Route<dynamic> route) => false);
-              },
+            Container(
+              margin: EdgeInsets.only(left: 15, right: 15, top: 10),
               child: ListTile(
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(100)),
+                  onTap: () async {
+                    if (await confirm(context)) {
+                      FirebaseAuth.instance.currentUser!.delete();
+                      Navigator.of(context).pushAndRemoveUntil(
+                          MaterialPageRoute(
+                            builder: (context) => SignInPage(
+                              change: widget.change,
+                            ),
+                          ),
+                          (Route<dynamic> route) => false);
+                    }
+                  },
                   leading: Icon(
                     Icons.delete,
                   ),
@@ -104,18 +114,23 @@ class _profileState extends State<profile> {
                     S.of(context).deleteAccount,
                   )),
             ),
-            GestureDetector(
-              onTap: () {
-                FirebaseAuth.instance.signOut();
-                Navigator.of(context).pushAndRemoveUntil(
-                    MaterialPageRoute(
-                      builder: (context) => SignInPage(
-                        change: widget.change,
-                      ),
-                    ),
-                    (Route<dynamic> route) => false);
-              },
+            Container(
+              margin: EdgeInsets.symmetric(horizontal: 15, vertical: 10),
               child: ListTile(
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(100)),
+                  onTap: () async {
+                    if (await confirm(context)) {
+                      FirebaseAuth.instance.signOut();
+                      Navigator.of(context).pushAndRemoveUntil(
+                          MaterialPageRoute(
+                            builder: (context) => SignInPage(
+                              change: widget.change,
+                            ),
+                          ),
+                          (Route<dynamic> route) => false);
+                    }
+                  },
                   leading: Icon(
                     Icons.logout,
                   ),
@@ -125,32 +140,52 @@ class _profileState extends State<profile> {
             ),
             Text(S.of(context).more),
             Container(
-                margin: EdgeInsets.all(7),
+                margin: EdgeInsets.only(left: 15, right: 15, top: 10),
                 child: ListTile(
+                    shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(100)),
                     leading: Icon(
                       Icons.question_mark_rounded,
                     ),
                     title: Text(
                       S.of(context).about,
                     ))),
-            ListTile(
-                leading: Icon(
-                  Icons.shield,
-                ),
-                title: Text(
-                  S.of(context).terms,
-                )),
-            ListTile(
-                leading: Icon(
-                  Icons.security,
-                ),
-                title: Text(
-                  S.of(context).privary,
-                )),
-            Text(S.of(context).language_name),
-            GestureDetector(
-              onTap: widget.change,
+            Container(
+              margin: EdgeInsets.only(left: 15, right: 15, top: 10),
               child: ListTile(
+                  onTap: () {
+                    Navigator.of(context).push(MaterialPageRoute(
+                      builder: (context) => MyWidget(),
+                    ));
+                  },
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(100)),
+                  leading: Icon(
+                    Icons.shield,
+                  ),
+                  title: Text(
+                    S.of(context).terms,
+                  )),
+            ),
+            Container(
+              margin: EdgeInsets.only(left: 15, right: 15, top: 10),
+              child: ListTile(
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(100)),
+                  leading: Icon(
+                    Icons.security,
+                  ),
+                  title: Text(
+                    S.of(context).privary,
+                  )),
+            ),
+            Text(S.of(context).language_name),
+            Container(
+              margin: EdgeInsets.symmetric(horizontal: 15, vertical: 10),
+              child: ListTile(
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(100)),
+                onTap: widget.change,
                 leading: Icon(
                   Icons.language,
                 ),
